// src/app/transfers/models/api/partnerModel.ts
/**
 * OMRO DigitalApp API
 * API used on OMRO Digital Resources
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface PartnerModel { 
    /**
     * 2 letter country code. RO for Romania
     */
    countryCode?: string;
    /**
     * Name of the partner
     */
    name?: string;
    /**
     * PF or PJ
     */
    partnerType?: string;
    /**
     * CUI for Romanian companies, Empty string for PF
     */
    uniqueCode?: string;
}